+++ glomar_gridding/utils.py
@@ -7,7 +7,6 @@ from datetime import date, timedelta
 from enum import IntEnum
 import inspect
 import os
-import subprocess
 import logging
 from typing import TypeVar
 import netCDF4 as nc
@@ -327,33 +326,51 @@ def get_pentad_range(centre_date: date) -> tuple[date, date]:
     return start_date, end_date
 
 
-def init_logging(file: str | None = None) -> None:
+def _get_logging_level(level: str) -> int:
+    match level.lower():
+        case "debug":
+            level_i = 10
+        case "info":
+            level_i = 20
+        case "warn":
+            level_i = 30
+        case "error":
+            level_i = 40
+        case "critical":
+            level_i = 50
+        case _:
+            raise ValueError(f"Unknown logging level: {level}")
+    return level_i
+
+
+def init_logging(file: str | None, level: str = "DEBUG") -> None:
     """
     Initialise the logger
 
     Parameters
     ----------
-    file : str | None
+    file : str
         File to send log messages to. If set to None (default) then print log
         messages to STDout
+    level : str
+        Level of logging, one of: "debug", "info", "warn", "error", "critical".
+
+    Returns
+    -------
+    None
     """
     from importlib import reload
 
-    reload(logging)  # Clear the logging from cdm_reader_mapper
-    if file is None:
-        logging.basicConfig(
-            format="\n%(levelname)s at %(asctime)s : %(message)s\n",
-            level=logging.INFO,
-        )
-    else:
-        logging.basicConfig(
-            filename=file,
-            filemode="a",
-            encoding="utf-8",
-            format="%(levelname)s at %(asctime)s : %(message)s",
-            level=logging.INFO,
-        )
+    level_i: int = _get_logging_level(level)
 
+    reload(logging)  # Clear the logging from cdm_reader_mapper
+    logging.basicConfig(
+        filename=file,
+        filemode="a",
+        encoding="utf-8",
+        format="%(levelname)s at %(asctime)s : %(message)s",
+        level=level_i,
+    )
     logging.captureWarnings(True)
     return None
 
@@ -380,10 +397,3 @@ def get_date_index(year: int, month: int, start_year: int) -> int:
         the first month of year `start_year`.
     """
     return 12 * (year - start_year) + (month - 1)
-
-
-def get_git_commit() -> str:
-    """Get the most recent commit of the repository for reproducibility"""
-    path = os.path.dirname(os.path.abspath(__file__))
-    out = subprocess.check_output(["git", "rev-parse", "HEAD"], cwd=path)
-    return out.strip().decode()