+++ CHANGES.md
@@ -6,11 +6,17 @@ Contributors to this version: Joseph Siddons (@josidd)
 
 ### New features and enhancements
 
+* `init_logging` now has a `level` argument (!50)
 * Added script to combine LSAT and SST for HadCRUT reconstruction using weights file using polars to join (!40)
 * Added `io.get_recurse` for scanning nested dictionaries by a key list (!38)
 
+### Bug fixes
+
+* Correct local import of `noc_helpers` in `noc_runners` scripts (!45)
+
 ### Breaking changes
 
+* `utils.get_git_commit` moved to `noc_helpers.get_git_commit`. Use of subprocess dropped (!44)
 * Refactored HadCRUT runner script (!38)
     * Outputs are no-longer yearly files
     * Use xarray rather than open a netCDF file